Added a proof of concept for testing JSON
One of the issues when working on the tests for get trustees is that we ended up
with a great deal of JSON within our test suite. We decided it'd be better if we
could have external JSON files with examples of good & bad JSON which we could
pull in and reference from our tests.

This branch proves that we can do that and gives an idea for a possible
implementation within our tests. When we refactor the get trustees spec we can
refine it further based on real world usage.